+++ R/with_progress.R
@@ -222,6 +222,8 @@ with_progress <- function(expr, handlers = progressr::handlers(), cleanup = TRUE
     
     ## Evaluate expression
     capture_conditions <- TRUE
+    finished <- FALSE
+    
     withCallingHandlers({
       res <- withVisible(expr)
     },
@@ -229,6 +231,19 @@ with_progress <- function(expr, handlers = progressr::handlers(), cleanup = TRUE
     progression = function(p) {
       progression_counter <<- progression_counter + 1
       if (debug) message(sprintf("- received a %s (n=%g)", sQuote(class(p)[1]), progression_counter))
+
+      if (finished) {
+        ## We might receive zero-amount progress updates after the fact that
+        ## the progress has been completed ...
+        amount <- p$amount
+        if (!is.numeric(amount) || amount > 0) {
+          ## ... but otherwise, it might be a coding mistake
+          msg <- conditionMessage(p)
+          if (length(msg) == 0) msg <- "character(0)"
+          type <- p$type
+          warning(sprintf("Received a progression %s request (amount=%g; msg=%s) but is not listening to this progressor. This can happen when code signals more progress updates than it configured the progressor to do. When the progressor completes all steps, it shuts down resulting in the global progression handler to no longer listen to it. To troubleshoot this, try with progressr::handlers(\"debug\")", sQuote(type), amount, sQuote(msg)))
+        }
+      }
       
       ## Don't capture conditions that are produced by progression handlers
       capture_conditions <<- FALSE
@@ -246,9 +261,8 @@ with_progress <- function(expr, handlers = progressr::handlers(), cleanup = TRUE
 
       ## Let the registered 'progressr' calling handlers process
       ## the 'progression' condition. If this resulted in the
-      ## progressor being completed, then 'finished' is TRUE,
-      ## which we do not make use of.
-      finished <- calling_handler(p)
+      ## progressor being completed, then 'finished' is TRUE
+      finished <<- calling_handler(p)
     },
   
     interrupt = handle_interrupt_or_error,
@@ -256,7 +270,7 @@ with_progress <- function(expr, handlers = progressr::handlers(), cleanup = TRUE
     error = handle_interrupt_or_error,
   
     condition = function(c) {
-      if (!capture_conditions) return()
+      if (!capture_conditions || finished) return()
       
       if (debug) message("- received a ", sQuote(class(c)[1]))